requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
contourpy==1.0.6
cycler==0.11.0
et-xmlfile==1.1.0
Flask==2.2.2
fonttools==4.38.0
frozenlist==1.3.3
idna==3.4
importlib-metadata==5.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.6.2
multidict==6.0.4
numpy==1.24.0
openai==0.25.0
openpyxl==3.0.10
packaging==22.0
pandas==1.5.2
pandas-stubs==1.5.2.221213
Pillow==9.3.0
plotly==5.11.0
pyee==9.0.4
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.7
requests==2.28.1
scikit-learn==1.2.0
scipy==1.9.3
six==1.16.0
slack-sdk==3.19.5
slackclient==2.9.4
slackeventsapi==3.0.1
tenacity==8.1.0
threadpoolctl==3.1.0
tqdm==4.64.1
types-pytz==2022.7.0.0
typing_extensions==4.4.0
urllib3==1.26.13
Werkzeug==2.2.2
yarl==1.8.2
zipp==3.11.0
